Taskqueue: move worker into queue module.
Move the queue worker from the worker.py into the queue.py so that the
enqueuing mechanism and the worker are in the same location for better
understanding of the code.
